Add a base class for the rolling classes.
rolling_mean.py, rolling_variance.py, rolling_skewness.py, and rolling_kurtosis.py inherit from rolling_base.py
