+++ pywatts/modules/__init__.py
@@ -7,6 +7,10 @@ from pywatts.modules.missing_value_detection import MissingValueDetector
 from pywatts.modules.profile_neural_network import ProfileNeuralNetwork
 from pywatts.modules.resample import Resampler
 from pywatts.modules.rolling_mean import RollingMean
+from pywatts.modules.rolling_kurtosis import RollingKurtosis
+from pywatts.modules.rolling_skewness import RollingSkewness
+from pywatts.modules.rolling_variance import RollingVariance
+from pywatts.modules.rolling_base import RollingGroupBy
 from pywatts.modules.root_mean_squared_error import RmseCalculator
 from pywatts.modules.sample_module import Sampler
 from pywatts.modules.trend_extraction import TrendExtraction
