+++ tests/unit/modules/test_rolling_skewness.py
@@ -1,25 +1,22 @@
 import unittest
 
 import pandas as pd
-import pytest
 import xarray as xr
-import numpy as np
 
-from pywatts.core.exceptions.wrong_parameter_exception import WrongParameterException
 from pywatts.modules.rolling_base import RollingGroupBy
-from pywatts.modules.rolling_variance import RollingVariance
+from pywatts.modules.rolling_skewness import RollingSkewness
 
 
-class TestRollingVariance(unittest.TestCase):
+class TestRollingSkewness(unittest.TestCase):
     def setUp(self) -> None:
-        self.rolling_variance = RollingVariance(window_size=3)
+        self.rolling_skewness = RollingSkewness(window_size=3)
 
     def tearDown(self) -> None:
-        self.rolling_variance = None
+        self.rolling_skewness = None
 
     def test_get_params(self):
         self.assertEqual(
-            self.rolling_variance.get_params(),
+            self.rolling_skewness.get_params(),
             {
                 "window_size": 3,
                 "window_size_unit": "d",
@@ -31,7 +28,7 @@ class TestRollingVariance(unittest.TestCase):
 
     def test_set_params(self):
         self.assertEqual(
-            self.rolling_variance.get_params(),
+            self.rolling_skewness.get_params(),
             {
                 "window_size": 3,
                 "window_size_unit": "d",
@@ -40,9 +37,9 @@ class TestRollingVariance(unittest.TestCase):
                 "continent": "Europe"
             }
         )
-        self.rolling_variance.set_params(window_size=5)
+        self.rolling_skewness.set_params(window_size=5)
         self.assertEqual(
-            self.rolling_variance.get_params(),
+            self.rolling_skewness.get_params(),
             {
                 "window_size": 5,
                 "window_size_unit": "d",
@@ -55,10 +52,10 @@ class TestRollingVariance(unittest.TestCase):
     def test_transform_groupbyNo(self):
         time = pd.date_range('2002-01-01', freq='24H', periods=7)
 
-        da = xr.DataArray([2, 3, 4, 5, 6, 7, 8], dims=["time"], coords={'time': time})
+        da = xr.DataArray([2, 2, 2, 5, 8, 8, 8], dims=["time"], coords={'time': time})
 
-        result = self.rolling_variance.transform(da)
+        result = self.rolling_skewness.transform(da)
 
-        expected_result = xr.DataArray([0., 0, 0.5, 1., 1., 1., 1.], dims=["time"], coords={'time': time})
+        expected_result = xr.DataArray([0., 0., 0., 0., 1.732051, 0., -1.732051], dims=["time"], coords={'time': time})
 
-        xr.testing.assert_equal(result, expected_result)
+        xr.testing.assert_allclose(result, expected_result)
